PJT/집SEEK/requirements.txt
ace_tools==0.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
fiona==1.10.1
fonttools==4.56.0
geopandas==1.0.1
idna==3.10
importlib_resources==6.5.2
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rtree==1.4.0
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.7
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
